requirements.txt
arcade==3.3.2
pillow>=10.4.0
